API to Python + tests

+++ runtime/host_cdts_converter.h
@@ -13,5 +13,5 @@
 extern "C"
 {
 	cdts* convert_host_params_to_cdts(PyObject* params_names, PyObject* params_types, metaffi_size return_values_size);
-	PyObject* convert_host_return_values_from_cdts(cdts* pcdts, int index);
+	PyObject* convert_host_return_values_from_cdts(cdts* pcdts, uint64_t index);
 }
